Improve C header file
This allows the C header file to be included in C++ applications as
well.

Furthermore, the strings that are not supposed to be modified are made
const.  This is for format and settings specifiers (which could be
string literals), as well as the writer callback which should just read
from the buffer.

Signed-off-by: Shahbaz Youssefi <ShabbyX@gmail.com>